refactor: extract screenshot into a page component
