+++ src/components/pages/home/index.js
@@ -1,3 +1,4 @@
-import Hero from './Hero'
+import Hero from './hero'
+import Screenshots from './screenshots'
 
-export { Hero }
+export { Hero, Screenshots }
