+++ src/pages/index.js
@@ -11,23 +11,15 @@ import {
   Container as ContainerBase,
   Flex,
   Heading,
-  Image,
-  Hide,
   Link,
   Subhead,
   Text,
   Card
 } from 'components/elements'
 
-import {
-  Grid,
-  Header,
-  Layout,
-  LiveDemo,
-  PricingTable
-} from 'components/patterns'
+import { Header, Layout, LiveDemo, PricingTable } from 'components/patterns'
 
-import { Hero } from 'components/pages/home'
+import { Hero, Screenshots } from 'components/pages/home'
 
 import { borders, colors } from 'theme'
 import get from 'dlv'
@@ -364,76 +356,6 @@ const Subheader = ({ title, subtitle, textAlign = 'center', children }) => (
     {children}
   </>
 )
-
-const Screenshots = props => {
-  const screenshotsUrls = [
-    { theme: 'dark', brand: 'Apple' },
-    { theme: 'light', brand: 'MDN' },
-    { theme: 'light', brand: 'StackOverflow' },
-    { theme: 'light', brand: 'ProductHunt' },
-    { theme: 'dark', brand: 'Nasa' }
-  ].map(item => {
-    const id = item.brand.toLowerCase()
-    const filename = `${id}.png`
-    return `https://cdn.microlink.io/website/browser/${item.theme}/${filename}`
-  })
-
-  const [screenshotUrl, setScreenshotUrl] = useState(screenshotsUrls[0])
-  const [index, setIndex] = useState(0)
-
-  const handleClick = event => {
-    event.preventDefault()
-    setIndex((index + 1) % screenshotsUrls.length)
-    setScreenshotUrl(screenshotsUrls[index])
-  }
-
-  const blockOne = (
-    <Flex justifyContent='center' alignItems='center' flexDirection='column'>
-      <Subheader subtitle='screenshot' title='Turn websites into a snapshot' />
-    </Flex>
-  )
-
-  const blockTwo = (
-    <Flex justifyContent='center' alignItems='center' flexDirection='column'>
-      <Box data-tilt pt={3}>
-        <Image src={screenshotUrl} />
-      </Box>
-      <Text
-        mt={[0, 0, 0, 3]}
-        mb={[0, 0, 0, 3]}
-        px={6}
-        textAlign='center'
-        maxWidth='960px'
-      >
-        Take a retina display screenshot of any URL. Export them to PNG or JPEG.
-        Automatic CDN redistribution. Overlay composition using a browser framer
-        & background. Just in time stale refresh, keeping them up to date.
-      </Text>
-      <Flex pt={3} alignItems='center' justifyContent='center'>
-        <Button onClick={() => navigate('/screenshot')}>
-          <Caps fontSize={0}>Live Demo</Caps>
-        </Button>
-        <Link
-          onClick={() => navigate('/docs/api/parameters/screenshot')}
-          ml={3}
-        >
-          Read Docs
-        </Link>
-      </Flex>
-    </Flex>
-  )
-
-  return (
-    <Block
-      id='screenshot'
-      flexDirection='column'
-      onClick={handleClick}
-      blockOne={blockOne}
-      blockTwo={blockTwo}
-      {...props}
-    />
-  )
-}
 
 const slide = keyframes`
 from {
